backend/requirements.txt
# FastAPI и основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# База данных
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP клиент для парсинга
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# Кеширование
redis==5.0.1
aioredis==2.0.1

# Планировщик задач
celery==5.3.4
celery[redis]==5.3.4

# Утилиты
python-dotenv==1.0.0
loguru==0.7.2
slowapi==0.1.9

# Для работы с изображениями
Pillow==10.1.0

# Валидация и форматирование
email-validator==2.1.0
